backend-django/requirements.txt
annotated-types==0.7.0
anyio==4.12.0
asgiref==3.10.0
brotli==1.2.0
certifi==2025.11.12
cffi==2.0.0
charset-normalizer==3.4.4
colorama==0.4.6
cssselect2==0.8.0
distro==1.9.0
Django==5.2.7
django-ckeditor==6.7.3
django-cors-headers==4.3.1
django-filter==25.2
django-js-asset==3.1.2
django-phonenumber-field==7.3.0
django-taggit==6.1.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
fonttools==4.61.1
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.11
jiter==0.12.0
numpy==2.3.5
openai==2.9.0
phonenumbers==8.13.26
pillow==12.0.0
psycopg2-binary==2.9.11
pycparser==2.23
pydantic==2.12.5
pydantic_core==2.41.5
pydyf==0.12.1
PyJWT==2.10.1
pyphen==0.17.2
python-decouple==3.8
pytz==2025.2
razorpay==2.0.0
requests==2.31.0
sniffio==1.3.1
sqlparse==0.5.3
swig==4.4.0
tinycss2==1.5.1
tinyhtml5==2.0.0
tqdm==4.67.1
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.5.0
weasyprint==67.0
webencodings==0.5.1
zopfli==0.4.0
celery==5.3.4
redis==5.0.1
django-celery-beat
django-celery-results
flower
gunicorn
firebase-admin
cryptography==46.0.5